Successfully built and installed the CE SDK
Haven't tried to run it, but everything built which is promising.
Notably, it turns out that LLVM development moved to a different
repository which already had fixes for the SmallVector error I patched,
and fixed a bug in instruction lowering that I bumped into when trying
to build the libraries.